wound-care-ai/backend/requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0

# Database
mysql-connector-python==8.2.0
SQLAlchemy==2.0.23
PyMySQL==1.1.0

# AI/ML Libraries
torch==2.1.1
torchvision==0.16.1
segmentation-models-pytorch==0.3.3
albumentations==1.3.1
timm==0.9.2
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
Pillow==10.1.0
scipy==1.11.4

# Image Processing
scikit-image==0.22.0

# PDF Generation
reportlab==4.0.7
PyPDF2==3.0.1

# API & Utilities
requests==2.31.0
python-multipart==0.0.6
Werkzeug==3.0.1

# Password Hashing
bcrypt==4.1.1

# Date/Time
python-dateutil==2.8.2

# AI Chatbot (OpenAI or local LLM)
openai==1.3.7

# WebSocket for real-time chat
flask-socketio==5.3.5
python-socketio==5.10.0

# Validation
marshmallow==3.20.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0

